Genservers API

Installation

GenServers.GenServers — Module

GenServers

Implements a generic server Actor's protocol.

The current stable, registered version is installed with

pkg> add Actors

The development version is installed with:

pkg> add "https://github.com/JuliaActors/Actors.jl"

GenServers.version — Constant

Gives the package version.

julia> using GenServers

julia> GenServers.version
v"0.1.0"

Starting a GenServer

GenServers.genserver — Function

genserver(m::Module, args...; name=nothing, 
    pid=myid(), thrd=false, 
    sticky=false, taskref=nothing)

Create an actor in :genserver mode. It uses the callbacks in a user provided module m when processing messages.

Arguments

m::Module: a user provided module in current scope,

args...: arguments to the user provided init callback.

Keyword Arguments

name=nothing: if a name::Symbol is provided the server is registered and the name is returned,

pid=myid(): worker pid to create the actor on,

thrd=false: thread to create the actor on,

sticky=false: if true, the actor is created in the same thread,

taskref=nothing: if a Ref{Task} variable is provided, it gets the created Task.